Fixed comments and move one class
