+++ src/main/java/net/sf/jabref/gui/BasePanel.java
@@ -55,10 +55,10 @@ import net.sf.jabref.gui.desktop.JabRefDesktop;
 import net.sf.jabref.gui.entryeditor.EntryEditor;
 import net.sf.jabref.gui.exporter.ExportToClipboardAction;
 import net.sf.jabref.gui.exporter.SaveDatabaseAction;
-import net.sf.jabref.gui.externalfiles.ExternalFileMenuItem;
 import net.sf.jabref.gui.externalfiles.FindFullTextAction;
 import net.sf.jabref.gui.externalfiles.SynchronizeFileField;
 import net.sf.jabref.gui.externalfiles.WriteXMPAction;
+import net.sf.jabref.gui.externalfiletype.ExternalFileMenuItem;
 import net.sf.jabref.gui.externalfiletype.ExternalFileType;
 import net.sf.jabref.gui.externalfiletype.ExternalFileTypes;
 import net.sf.jabref.gui.fieldeditors.FieldEditor;
