+++ src/main/java/net/sf/jabref/gui/externalfiletype/ExternalFileTypes.java
@@ -317,9 +317,6 @@ public final class ExternalFileTypes {
         }
 
         // Finally, build the list of types based on the modified defaults list:
-        for (ExternalFileType type : types) {
-            externalFileTypes.add(type);
-        }
+        externalFileTypes.addAll(types);
     }
-
 }
